requirements.txt
streamlit
google-generativeai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
uvicorn
FastAPI